Change scale scheme of skeletons on export
Since GR2 uses centimeters and MDB uses meters, when converting NWN2 to FBX, the tool set the scale of skeletons to 0.01. From now, the bones are converted from centimeters to meters and the scale of skeletons is set to 1. I think this scheme is easier for artists to work with.